+++ CHANGELOG.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.11.0] - Unreleased
+## Unreleased
+
+### Added
+
+- Added Strip.text property https://github.com/Textualize/textual/issues/1620
 
 ### Changed
 
 - The default filename for screenshots uses a datetime format similar to ISO8601, but with reserved characters replaced by underscores https://github.com/Textualize/textual/pull/1518
+
 
 ## [0.10.0] - 2023-01-19